Fix: Aligned example and maplibre_gl_web pubspec versions. (#476)
Hi guys, I was trying to run the example application and I noticed that
something was going wrong.
Analyzing the problem, I found that the versions of some plugins had not
been updated since the new 0.20.0 update in the main branch.

With this PR I updated the versioning of the pubspec.yaml files of
example and maplibre_gl_web.

I know it's a simple solution and PR, but useful, also because without
it, it isn't possible to run pub gets and the code isn't aligned between
folders. Also, it isn't possible to run the example application.